Implemented AccountFactory
Refactored most of the objects related to accounts, such admin and user. Created SysadminClass, which is the only user that is allowed to create admins. 
Renamed methods within userManagement and ArticleManagement to conform to the CRUD methodology. Implemented the Factory Pattern for the creation of the account objects using createAccount method.